+++ database/migrations/2025_03_26_143954_create_payments_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('student_id');
             $table->enum('payment_style', ['current', 'due']);
             $table->integer('due_date')->default(7);
-            $table->decimal('tuition_fee', 10, 2);
+            $table->integer('tuition_fee');
             $table->timestamps();
         });
     }
